+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -1,6 +1,8 @@
 package wooteco.subway.line.application;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.line.application.exception.DuplicateLineException;
+import wooteco.subway.line.application.exception.NoLineException;
 import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.line.domain.Line;
