+++ src/main/java/wooteco/subway/line/application/exception/OverDistanceOfSectionException.java
@@ -5,4 +5,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.BAD_REQUEST, reason = "추가할 구간의 길이가 기존 구간의 길이보다 클 수 없습니다.")
 public class OverDistanceOfSectionException extends RuntimeException {
+    public OverDistanceOfSectionException() {
+    }
 }
